move Header from Container to separate component
