+++ src/features/MoviesPage/index.js
@@ -3,7 +3,8 @@ import { useSelector, useDispatch } from "react-redux";
 import Section from "../../common/Section";
 import Pagination from "./../../common/Pagination/index";
 import { StatusChecker } from "./../../common/StatusChecker/index";
-import { Container, Header } from "./../../common/Container/index";
+import { Container } from "./../../common/Container/index";
+import { Header } from "./../../common/Header/index";
 import { NoResult } from "./../../common/NoResult";
 import { pageState } from "./../../common/pageState";
 import { selectGenerateList } from "./../../common/commonSlice";
