+++ src/features/PeoplePage/index.js
@@ -1,6 +1,7 @@
 import { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { Container, Header } from "../../common/Container";
+import { Container } from "../../common/Container";
+import { Header } from "../../common/Header";
 import Section from "../../common/Section";
 import Pagination from "./../../common/Pagination/index";
 import { StatusChecker } from "./../../common/StatusChecker/index";
